Started moving delegation queries into own module
