+++ contracts/mixnet/src/contract.rs
@@ -4,13 +4,14 @@
 use crate::gateways::queries::query_gateways_paged;
 use crate::gateways::queries::query_owns_gateway;
 use crate::helpers::calculate_epoch_reward_rate;
+use crate::mixnet_params::queries::query_state_params;
 use crate::mixnet_params::state::State;
 use crate::mixnodes::bonding_queries as mixnode_queries;
-use crate::mixnodes::bonding_queries::query_all_mixnode_delegations_paged;
-use crate::mixnodes::bonding_queries::query_mixnode_delegation;
 use crate::mixnodes::bonding_queries::query_mixnode_delegations_paged;
 use crate::mixnodes::bonding_queries::query_mixnodes_paged;
-use crate::mixnodes::bonding_queries::query_reverse_mixnode_delegations_paged;
+use crate::mixnodes::delegation_queries::query_all_mixnode_delegations_paged;
+use crate::mixnodes::delegation_queries::query_mixnode_delegation;
+use crate::mixnodes::delegation_queries::query_reverse_mixnode_delegations_paged;
 use crate::storage::{config, layer_distribution};
 use crate::{error::ContractError, queries};
 use config::defaults::REWARDING_VALIDATOR_ADDRESS;
@@ -186,7 +187,7 @@ pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> Result<QueryResponse, Cont
             to_binary(&mixnode_queries::query_owns_mixnode(deps, address)?)
         }
         QueryMsg::OwnsGateway { address } => to_binary(&query_owns_gateway(deps, address)?),
-        QueryMsg::StateParams {} => to_binary(&queries::query_state_params(deps)),
+        QueryMsg::StateParams {} => to_binary(&query_state_params(deps)),
         QueryMsg::CurrentRewardingInterval {} => {
             to_binary(&queries::query_rewarding_interval(deps))
         }
