+++ contracts/mixnet/src/mixnet_params/mod.rs
@@ -1,2 +1,3 @@
+pub mod queries;
 pub mod state;
 pub mod transactions;
