+++ contracts/mixnet/src/mixnodes/delegation_queries.rs
@@ -1,0 +1,381 @@
+use crate::error::ContractError;
+use crate::helpers::get_all_delegations_paged;
+use crate::queries::calculate_start_value;
+use crate::queries::DELEGATION_PAGE_DEFAULT_LIMIT;
+use crate::queries::DELEGATION_PAGE_MAX_LIMIT;
+use crate::storage::all_mix_delegations_read;
+use crate::storage::mix_delegations_read;
+use crate::storage::reverse_mix_delegations_read;
+use config::defaults::DENOM;
+use cosmwasm_std::coin;
+use cosmwasm_std::Addr;
+use cosmwasm_std::Deps;
+use cosmwasm_std::Order;
+use cosmwasm_std::StdResult;
+use mixnet_contract::Delegation;
+use mixnet_contract::IdentityKey;
+use mixnet_contract::PagedAllDelegationsResponse;
+use mixnet_contract::PagedReverseMixDelegationsResponse;
+use mixnet_contract::RawDelegationData;
+
+pub(crate) fn query_all_mixnode_delegations_paged(
+    deps: Deps,
+    start_after: Option<Vec<u8>>,
+    limit: Option<u32>,
+) -> StdResult<PagedAllDelegationsResponse<RawDelegationData>> {
+    let limit = limit
+        .unwrap_or(DELEGATION_PAGE_DEFAULT_LIMIT)
+        .min(DELEGATION_PAGE_MAX_LIMIT) as usize;
+
+    let bucket = all_mix_delegations_read::<RawDelegationData>(deps.storage);
+    let start = start_after.map(|mut v| {
+        v.push(0);
+        v
+    });
+    get_all_delegations_paged::<RawDelegationData>(&bucket, &start, limit)
+}
+
+pub(crate) fn query_reverse_mixnode_delegations_paged(
+    deps: Deps,
+    delegation_owner: Addr,
+    start_after: Option<IdentityKey>,
+    limit: Option<u32>,
+) -> StdResult<PagedReverseMixDelegationsResponse> {
+    let limit = limit
+        .unwrap_or(DELEGATION_PAGE_DEFAULT_LIMIT)
+        .min(DELEGATION_PAGE_MAX_LIMIT) as usize;
+    let start = calculate_start_value(start_after);
+
+    let delegations = reverse_mix_delegations_read(deps.storage, &delegation_owner)
+        .range(start.as_deref(), None, Order::Ascending)
+        .take(limit)
+        .map(|res| {
+            res.map(|entry| {
+                String::from_utf8(entry.0)
+                    .expect("Non-UTF8 address used as key in bucket. The storage is corrupted!")
+            })
+        })
+        .collect::<StdResult<Vec<IdentityKey>>>()?;
+
+    let start_next_after = delegations.last().cloned();
+
+    Ok(PagedReverseMixDelegationsResponse::new(
+        delegation_owner,
+        delegations,
+        start_next_after,
+    ))
+}
+
+// queries for delegation value of given address for particular node
+pub(crate) fn query_mixnode_delegation(
+    deps: Deps,
+    mix_identity: IdentityKey,
+    address: Addr,
+) -> Result<Delegation, ContractError> {
+    match mix_delegations_read(deps.storage, &mix_identity).may_load(address.as_bytes())? {
+        Some(delegation_value) => Ok(Delegation::new(
+            address,
+            coin(delegation_value.amount.u128(), DENOM),
+            delegation_value.block_height,
+        )),
+        None => Err(ContractError::NoMixnodeDelegationFound {
+            identity: mix_identity,
+            address,
+        }),
+    }
+}
+
+#[cfg(test)]
+pub(crate) mod tests {
+    use super::*;
+    use crate::storage::{mix_delegations};
+    use crate::support::tests::helpers;
+    use crate::support::tests::helpers::raw_delegation_fixture;
+    
+    
+    use cosmwasm_std::{Addr, Storage};
+    
+    
+    
+
+    pub fn store_n_mix_delegations(n: u32, storage: &mut dyn Storage, node_identity: &str) {
+        for i in 0..n {
+            let address = format!("address{}", i);
+            mix_delegations(storage, node_identity)
+                .save(address.as_bytes(), &raw_delegation_fixture(42))
+                .unwrap();
+        }
+    }
+
+    #[cfg(test)]
+    mod querying_for_mixnode_delegations_paged {
+        use super::*;
+        use crate::mixnodes::bonding_queries::query_mixnode_delegations_paged;
+        use mixnet_contract::IdentityKey;
+
+        #[test]
+        fn retrieval_obeys_limits() {
+            let mut deps = helpers::init_contract();
+            let limit = 2;
+            let node_identity: IdentityKey = "foo".into();
+            store_n_mix_delegations(100, &mut deps.storage, &node_identity);
+
+            let page1 = query_mixnode_delegations_paged(
+                deps.as_ref(),
+                node_identity,
+                None,
+                Option::from(limit),
+            )
+            .unwrap();
+            assert_eq!(limit, page1.delegations.len() as u32);
+        }
+
+        #[test]
+        fn retrieval_has_default_limit() {
+            let mut deps = helpers::init_contract();
+            let node_identity: IdentityKey = "foo".into();
+            store_n_mix_delegations(
+                DELEGATION_PAGE_DEFAULT_LIMIT * 10,
+                &mut deps.storage,
+                &node_identity,
+            );
+
+            // query without explicitly setting a limit
+            let page1 =
+                query_mixnode_delegations_paged(deps.as_ref(), node_identity, None, None).unwrap();
+            assert_eq!(
+                DELEGATION_PAGE_DEFAULT_LIMIT,
+                page1.delegations.len() as u32
+            );
+        }
+
+        #[test]
+        fn retrieval_has_max_limit() {
+            let mut deps = helpers::init_contract();
+            let node_identity: IdentityKey = "foo".into();
+            store_n_mix_delegations(
+                DELEGATION_PAGE_DEFAULT_LIMIT * 10,
+                &mut deps.storage,
+                &node_identity,
+            );
+
+            // query with a crazily high limit in an attempt to use too many resources
+            let crazy_limit = 1000 * DELEGATION_PAGE_DEFAULT_LIMIT;
+            let page1 = query_mixnode_delegations_paged(
+                deps.as_ref(),
+                node_identity,
+                None,
+                Option::from(crazy_limit),
+            )
+            .unwrap();
+
+            // we default to a decent sized upper bound instead
+            let expected_limit = DELEGATION_PAGE_MAX_LIMIT;
+            assert_eq!(expected_limit, page1.delegations.len() as u32);
+        }
+
+        #[test]
+        fn pagination_works() {
+            let mut deps = helpers::init_contract();
+            let node_identity: IdentityKey = "foo".into();
+
+            mix_delegations(&mut deps.storage, &node_identity)
+                .save("1".as_bytes(), &raw_delegation_fixture(42))
+                .unwrap();
+
+            let per_page = 2;
+            let page1 = query_mixnode_delegations_paged(
+                deps.as_ref(),
+                node_identity.clone(),
+                None,
+                Option::from(per_page),
+            )
+            .unwrap();
+
+            // page should have 1 result on it
+            assert_eq!(1, page1.delegations.len());
+
+            // save another
+            mix_delegations(&mut deps.storage, &node_identity)
+                .save("2".as_bytes(), &raw_delegation_fixture(42))
+                .unwrap();
+
+            // page1 should have 2 results on it
+            let page1 = query_mixnode_delegations_paged(
+                deps.as_ref(),
+                node_identity.clone(),
+                None,
+                Option::from(per_page),
+            )
+            .unwrap();
+            assert_eq!(2, page1.delegations.len());
+
+            mix_delegations(&mut deps.storage, &node_identity)
+                .save("3".as_bytes(), &raw_delegation_fixture(42))
+                .unwrap();
+
+            // page1 still has 2 results
+            let page1 = query_mixnode_delegations_paged(
+                deps.as_ref(),
+                node_identity.clone(),
+                None,
+                Option::from(per_page),
+            )
+            .unwrap();
+            assert_eq!(2, page1.delegations.len());
+
+            // retrieving the next page should start after the last key on this page
+            let start_after = Addr::unchecked("2");
+            let page2 = query_mixnode_delegations_paged(
+                deps.as_ref(),
+                node_identity.clone(),
+                Option::from(start_after),
+                Option::from(per_page),
+            )
+            .unwrap();
+
+            assert_eq!(1, page2.delegations.len());
+
+            // save another one
+            mix_delegations(&mut deps.storage, &node_identity)
+                .save("4".as_bytes(), &raw_delegation_fixture(42))
+                .unwrap();
+
+            let start_after = Addr::unchecked("2");
+            let page2 = query_mixnode_delegations_paged(
+                deps.as_ref(),
+                node_identity,
+                Option::from(start_after),
+                Option::from(per_page),
+            )
+            .unwrap();
+
+            // now we have 2 pages, with 2 results on the second page
+            assert_eq!(2, page2.delegations.len());
+        }
+    }
+
+    #[cfg(test)]
+    mod querying_for_all_mixnode_delegations_paged {
+        use super::*;
+        use crate::helpers::identity_and_owner_to_bytes;
+        use mixnet_contract::IdentityKey;
+
+        #[test]
+        fn retrieval_obeys_limits() {
+            let mut deps = helpers::init_contract();
+            let limit = 2;
+            let node_identity: IdentityKey = "foo".into();
+            store_n_mix_delegations(100, &mut deps.storage, &node_identity);
+
+            let page1 =
+                query_all_mixnode_delegations_paged(deps.as_ref(), None, Option::from(limit))
+                    .unwrap();
+            assert_eq!(limit, page1.delegations.len() as u32);
+        }
+
+        #[test]
+        fn retrieval_has_default_limit() {
+            let mut deps = helpers::init_contract();
+            let node_identity: IdentityKey = "foo".into();
+            store_n_mix_delegations(
+                DELEGATION_PAGE_DEFAULT_LIMIT * 10,
+                &mut deps.storage,
+                &node_identity,
+            );
+
+            // query without explicitly setting a limit
+            let page1 = query_all_mixnode_delegations_paged(deps.as_ref(), None, None).unwrap();
+            assert_eq!(
+                DELEGATION_PAGE_DEFAULT_LIMIT,
+                page1.delegations.len() as u32
+            );
+        }
+
+        #[test]
+        fn retrieval_has_max_limit() {
+            let mut deps = helpers::init_contract();
+            let node_identity: IdentityKey = "foo".into();
+            store_n_mix_delegations(
+                DELEGATION_PAGE_DEFAULT_LIMIT * 10,
+                &mut deps.storage,
+                &node_identity,
+            );
+
+            // query with a crazily high limit in an attempt to use too many resources
+            let crazy_limit = 1000 * DELEGATION_PAGE_DEFAULT_LIMIT;
+            let page1 =
+                query_all_mixnode_delegations_paged(deps.as_ref(), None, Option::from(crazy_limit))
+                    .unwrap();
+
+            // we default to a decent sized upper bound instead
+            let expected_limit = DELEGATION_PAGE_MAX_LIMIT;
+            assert_eq!(expected_limit, page1.delegations.len() as u32);
+        }
+
+        #[test]
+        fn pagination_works() {
+            let mut deps = helpers::init_contract();
+            let node_identity: IdentityKey = "foo".into();
+
+            mix_delegations(&mut deps.storage, &node_identity)
+                .save("1".as_bytes(), &raw_delegation_fixture(42))
+                .unwrap();
+
+            let per_page = 2;
+            let page1 =
+                query_all_mixnode_delegations_paged(deps.as_ref(), None, Option::from(per_page))
+                    .unwrap();
+
+            // page should have 1 result on it
+            assert_eq!(1, page1.delegations.len());
+
+            // save another
+            mix_delegations(&mut deps.storage, &node_identity)
+                .save("2".as_bytes(), &raw_delegation_fixture(42))
+                .unwrap();
+
+            // page1 should have 2 results on it
+            let page1 =
+                query_all_mixnode_delegations_paged(deps.as_ref(), None, Option::from(per_page))
+                    .unwrap();
+            assert_eq!(2, page1.delegations.len());
+
+            mix_delegations(&mut deps.storage, &node_identity)
+                .save("3".as_bytes(), &raw_delegation_fixture(42))
+                .unwrap();
+
+            // page1 still has 2 results
+            let page1 =
+                query_all_mixnode_delegations_paged(deps.as_ref(), None, Option::from(per_page))
+                    .unwrap();
+            assert_eq!(2, page1.delegations.len());
+
+            // retrieving the next page should start after the last key on this page
+            let start_after = identity_and_owner_to_bytes(&node_identity, &Addr::unchecked("2"));
+            let page2 = query_all_mixnode_delegations_paged(
+                deps.as_ref(),
+                Option::from(start_after.clone()),
+                Option::from(per_page),
+            )
+            .unwrap();
+
+            assert_eq!(1, page2.delegations.len());
+
+            // save another one
+            mix_delegations(&mut deps.storage, &node_identity)
+                .save("4".as_bytes(), &raw_delegation_fixture(42))
+                .unwrap();
+
+            let page2 = query_all_mixnode_delegations_paged(
+                deps.as_ref(),
+                Option::from(start_after),
+                Option::from(per_page),
+            )
+            .unwrap();
+
+            // now we have 2 pages, with 2 results on the second page
+            assert_eq!(2, page2.delegations.len());
+        }
+    }
+}
