+++ contracts/mixnet/src/mixnodes/mod.rs
@@ -2,3 +2,4 @@ pub mod bonding_queries;
 pub mod bonding_transactions;
 pub mod delegation_queries;
 pub mod delegation_transactions;
+pub mod layer_queries;
